requirements.txt
python-aalib>=0.3.2
requests>=2.4.3
Pillow>=2.6.1
ansicolors
